Fix render level passing bug
Absolute level was previously being worked out (on line 265, using the --render_level flag), but not being passed to RenderRegionsRandomColour.

Instead, FLAGS_render_level (in the range [0..1]) was being passed, which is the wrong value to use (unless --render_level was set to 0). 

This commit fixes this issue by passing absolute_level rather than FLAGS_render_level to RenderRegionsRandomColor.